requirements.txt
Flask==2.3.2
langchain[openai]==0.0.148
PyPDF2==3.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
openai>=0.27.0
numpy==1.23.5
tqdm==4.64.1
pydantic==1.10.7
requests==2.31.0
typing_extensions==4.5.0
tiktoken>=0.4.0
gunicorn==20.1.0

